+++ notification/src/main/java/com/microservice/erp/domain/repositories/INoticeConfigurationRepository.java
@@ -3,12 +3,13 @@ package com.microservice.erp.domain.repositories;
 import com.microservice.erp.domain.entities.NoticeConfiguration;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.CrossOrigin;
+
+import java.math.BigInteger;
+
 /**
  * @author Passang Lhamo
- *
  */
 
 @Repository
-public interface INoticeConfigurationRepository extends JpaRepository<NoticeConfiguration,Long> {
+public interface INoticeConfigurationRepository extends JpaRepository<NoticeConfiguration, BigInteger> {
 }
