+++ training-management/src/main/java/com/microservice/erp/domain/repository/IEnlistmentScheduleRepository.java
@@ -4,8 +4,8 @@ import com.microservice.erp.domain.entities.EnlistmentSchedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-import org.springframework.web.bind.annotation.CrossOrigin;
 
+import java.math.BigInteger;
 import java.util.Date;
 import java.util.List;
 
@@ -14,11 +14,11 @@ import java.util.List;
  */
 
 @Repository
-public interface IEnlistmentScheduleRepository extends JpaRepository<EnlistmentSchedule, Long> {
+public interface IEnlistmentScheduleRepository extends JpaRepository<EnlistmentSchedule, BigInteger> {
     @Query(value = "FROM tms_enlistment_schedule t " +
             "WHERE ((:fromDate IS NULL) OR (t.fromDate>=:fromDate) )" +
             "AND ((:status IS NULL) OR (t.status=:status) ) ")
-    List<EnlistmentSchedule> getAllByFromDate(Date fromDate,String status);
+    List<EnlistmentSchedule> getAllByFromDate(Date fromDate, String status);
 
     boolean existsByToDateAfter(Date fromDate);
 }
