+++ training-management/src/main/java/com/microservice/erp/domain/repository/IMedicalQuestionSubCategoryRepository.java
@@ -4,11 +4,13 @@ import com.microservice.erp.domain.entities.MedicalQuestionSubCategory;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 /**
  * @author Rajib Kumer Ghosh
  *
  */
 
 @Repository
-public interface IMedicalQuestionSubCategoryRepository extends JpaRepository<MedicalQuestionSubCategory, Integer> {
+public interface IMedicalQuestionSubCategoryRepository extends JpaRepository<MedicalQuestionSubCategory, BigInteger> {
 }
