+++ training-management/src/main/java/com/microservice/erp/services/iServices/IReadEnlistmentScheduleService.java
@@ -2,10 +2,11 @@ package com.microservice.erp.services.iServices;
 
 import com.microservice.erp.domain.entities.EnlistmentSchedule;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IReadEnlistmentScheduleService {
     List<EnlistmentSchedule> getAllEnlistmentScheduleList();
 
-    EnlistmentSchedule getEnlistmentScheduleById(Long id);
+    EnlistmentSchedule getEnlistmentScheduleById(BigInteger id);
 }
