+++ training-management/src/main/java/com/microservice/erp/services/iServices/IReadMedicalCategoryService.java
@@ -2,10 +2,11 @@ package com.microservice.erp.services.iServices;
 
 import com.microservice.erp.domain.entities.MedicalQuestionCategory;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IReadMedicalCategoryService {
-    MedicalQuestionCategory getAllMedicalCategoriesById(Long id);
+    MedicalQuestionCategory getAllMedicalCategoriesById(BigInteger id);
 
     List<MedicalQuestionCategory> getAllMedicalCategoryList();
     List<MedicalQuestionCategory> getAllActiveMedicalCatList();
