+++ training-management/src/main/java/com/microservice/erp/services/iServices/IReadMedicalQuestionService.java
@@ -2,10 +2,11 @@ package com.microservice.erp.services.iServices;
 
 import com.microservice.erp.domain.entities.MedicalQuestionnaire;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface IReadMedicalQuestionService {
     List<MedicalQuestionnaire> getAllMedicalQuestionnaireList();
 
-    MedicalQuestionnaire getAllMedicalQuestionnaireById(Long id);
+    MedicalQuestionnaire getAllMedicalQuestionnaireById(BigInteger id);
 }
